Bug Fixing | Focusing
Fixed issue where the player would not be able to exit a focus station if they moved out of the area as they entered it. Fixed issue where the player camera kept moving close to the ground. Fixed issue where the player would display on the game camera. (Note: Made player invisible, however we could put the player on a layer that's not rendered by the player camera).