Player Movement Based on Joystick
Edited the player movement script to use the position of the joystick to determine input, rather than the input of the keyboard.
Created a math extensions script, extending the float class, to make this possible. Created a function to convert a value of a range to a value of a new range.