Refactored Audio Event Instances
Created a system to stop looping tracks automatically when starting a new track. Moderate refactor on the Audio Manager class by creating structs containing instances and references, rather than using the references/instances directly. This is done to facilitate the system of 'priority', which refers to stopping tracks of a matching priority when starting said track.